requirements.txt
ipykernel
scikit-learn

lightning
pyro-ppl
cmake
pykeops
gpytorch
numpy
pandas
scipy
seaborn
jupyter
biopython
pyfaidx
xlrd
bioservices
wordcloud
torchsummary
gcsfs
h5py

dmslogo==0.6.2
palettable==3.3.0
imageio==2.9.0
imageio-ffmpeg==0.4.4

arviz
tqdm

google-cloud-aiplatform
cloudml-hypertune

google-cloud-storage
